Assignment5Q1.js
In this sketch, I created a sketch that adds 5 bubbles that jitter randomly to a canvas of size 960 x 540. I used constructor(), move()/update(), and 
show() methods. Also I passed parameters for the x and y position when a new bubble is created, as well as passing the bubble radius.

Link: https://editor.p5js.org/Jake_Kingdon02/sketches/EQjARmcN2