Assignment5Q5.js
In this sketch, I created a sketch that animates my catchers so they will start at a random position and speed at the right edge of the screen, 
switching direction when they hit a canvas edge. I also detected any collision between a catcher and a breeder, and remove the breeder from the simulation.